+++ aiobungie/__init__.pyi
@@ -1,5 +1,6 @@
 from typing import Any
 
+from aiobungie import builders as builders
 from aiobungie import crate as crate
 from aiobungie import interfaces as interfaces
 from aiobungie import traits as traits
